+++ scripts/vscode.sh
@@ -1,4 +1,8 @@
-curl -L 'https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64' -o vscode.deb
-sudo dpkg -i vscode.deb
-sudo apt update && sudo apt install -y code
-unlink vscode.deb
+if ! command_exists code; then
+  echo installing vscode
+  curl -L 'https://code.visualstudio.com/sha/download?build=stable&os=linux-deb-x64' -o /tmp/vscode.deb
+  sudo dpkg -i /tmp/vscode.deb
+  sudo apt update && sudo apt install -y code
+  unlink /tmp/vscode.deb
+  echo vscode installed
+fi
